Move create button and update UI
